test(frontend): fix component error in `MkChart` story
